Added endpoints, added exceptions, updated my code

+++ Rate-My-Hike-BackEnd/src/main/java/com/example/ratemyhike/Service/EventServiceImpl.java
@@ -1,5 +1,7 @@
 package com.example.ratemyhike.Service;
 
+import com.example.ratemyhike.Exceptions.EventWithIdAlreadyExists;
+import com.example.ratemyhike.Exceptions.EventWithIdDoesntExist;
 import com.example.ratemyhike.Model.Event;
 import com.example.ratemyhike.Repo.EventRepo;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,25 +22,31 @@ public class EventServiceImpl implements EventService{
     }
 
     @Override
-    public Event createNewEvent(Event event) {
-        return eventRepo.save(event);
+    public Event createNewEvent(Event event) throws EventWithIdAlreadyExists {
+        if(!eventRepo.existsById(event.getEventNumber())) return eventRepo.save(event);
+        else throw new EventWithIdAlreadyExists();
     }
 
     @Override
-    public Event getEventById(Long id) {
-        return eventRepo.findEventByEventNumber(id).get();
+    public Event getEventById(Long id) throws EventWithIdDoesntExist {
+        if(eventRepo.existsById(id)) return eventRepo.findEventByEventNumber(id).get();
+        else throw new EventWithIdDoesntExist();
     }
 
     @Override
-    public void deleteEvent(Event event){
-        eventRepo.delete(event);
+    public void deleteEvent(Event event) throws EventWithIdDoesntExist{
+        if(!eventRepo.existsById(event.getEventNumber())) eventRepo.delete(event);
+        else throw new EventWithIdDoesntExist();
     }
 
     @Override
-    public void updateEvent(Long id, Event event) {
+    public void updateEvent(Long id, Event event) throws EventWithIdDoesntExist{
+        if(eventRepo.existsById(id)){
         Event oldEvent = eventRepo.findEventByEventNumber(id).get();
         event.setEventNumber(oldEvent.getEventNumber());
-        eventRepo.save(event);
+        eventRepo.save(event);}
+
+        else throw new EventWithIdDoesntExist();
     }
